# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Email Service API
  version: 1.0.0
  description: |
    A transactional email service for sending templated emails with support for custom branding.

    Features:
    - Send single or bulk emails
    - Template-based email rendering with variable substitution
    - Custom branding configuration per product
    - Idempotent email sending with idempotency keys
    - Email status tracking with delivery status updates
    - Async processing with queue-based delivery

    ## Authentication
    All API endpoints require authentication via API key. The key can be provided in two ways:
    - `X-API-Key` header
    - `Authorization: Bearer <api-key>` header
servers:
  - url: ""
    description: Current Origin (auto-detected from browser URL)
  - url: https://email.example.com
    description: Production
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (API key as token)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "invalid request body"
      required:
        - error
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns/{tag}/stats:
    get:
      tags:
        - Campaigns
      summary: Get campaign stats
      description: Returns aggregate send/delivery stats for a campaign tag.
      operationId: getCampaignStats
      parameters:
        - name: tag
          in: path
          required: true
          description: Campaign tag identifier
          schema:
            type: string
          example: "camp_migration_w1"
      responses:
        '200':
          description: Campaign stats
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing campaign tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/campaigns/{tag}/bounced:
    get:
      tags:
        - Campaigns
      summary: Get bounced emails for campaign
      description: Returns email addresses from a campaign that bounced or received complaints.
      operationId: getCampaignBounced
      parameters:
        - name: tag
          in: path
          required: true
          description: Campaign tag identifier
          schema:
            type: string
          example: "camp_migration_w1"
      responses:
        '200':
          description: List of bounced email addresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_tag:
                    type: string
                    example: "camp_migration_w1"
                  count:
                    type: integer
                    example: 5
                  addresses:
                    type: array
                    items:
                      type: string
                      format: email
                    example: ["bounced@example.com"]
        '400':
          description: Missing campaign tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/campaigns/{tag}/non-openers:
    get:
      tags:
        - Campaigns
      summary: Get non-opener emails for campaign
      description: Returns email addresses from a campaign that have not opened the email.
      operationId: getCampaignNonOpeners
      parameters:
        - name: tag
          in: path
          required: true
          description: Campaign tag identifier
          schema:
            type: string
          example: "camp_migration_w1"
      responses:
        '200':
          description: List of non-opener email addresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_tag:
                    type: string
                    example: "camp_migration_w1"
                  count:
                    type: integer
                    example: 42
                  addresses:
                    type: array
                    items:
                      type: string
                      format: email
        '400':
          description: Missing campaign tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/suppressions/check:
    post:
      tags:
        - Suppressions
      summary: Check suppressed emails
      description: |
        Given a list of email addresses, returns which ones are in the suppression list.
        Maximum 10,000 emails per request.
      operationId: checkSuppressions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emails
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                  maxItems: 10000
                  description: List of emails to check
                  example: ["user1@example.com", "bounced@example.com"]
      responses:
        '200':
          description: Suppression check results
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Number of suppressed emails found
                    example: 1
                  suppressed:
                    type: array
                    items:
                      type: string
                      format: email
                    description: Email addresses that are suppressed
                    example: ["bounced@example.com"]
        '400':
          description: Invalid request or too many emails
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "too many emails, max 10000"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/suppressions/{email}:
    delete:
      tags:
        - Suppressions
      summary: Remove email from suppression list
      description: |
        Removes an email address from the suppression list, allowing future
        emails to be sent to it. Use this to re-enable delivery after a user
        was unsubscribed, bounced, or complained.
      operationId: removeSuppression
      parameters:
        - name: email
          in: path
          required: true
          description: Email address to remove from suppression list
          schema:
            type: string
            format: email
          example: "user@example.com"
      responses:
        '200':
          description: Email removed from suppression list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email removed from suppression list"
                  email:
                    type: string
                    format: email
                    example: "user@example.com"
        '400':
          description: Missing email parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "email is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'